Create MHE FOPDT Simulator.py
The NewCar package is vehicle_model_v04.py. The point of this simulator is to try and model the car with an FOPDT relationship between the accelerator and the velocity. So far, this is only an MHE. It has troubles matching the predicted velocity to the actual value, but matches the engine torque well.